First push for drag and drop
Here is the first push for the drag and drop now that it is working and not just breaking everything. You can now pick up the card and move it around. The card won't drop in other places though, it will only go back to it's origin position. It's a little wonky how far you can move the card too, but still working on it